// boilerplate/src/payload-types.ts
export interface FeaturesBlock {
  sectionTitle?: string | null;
  sectionDescription?: string | null;
  features?:
    | {
        icon: 'calendar' | 'book' | 'users' | 'star' | 'code' | 'globe' | 'heart' | 'lightning';
        title: string;
        description: string;
        linkUrl?: string | null;
        linkLabel?: string | null;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'features';
}